refactor: removing store/tasks for now.
The task design was mainly for the purpose of allowing the web
to display progress, show previous builds/provisons/etc. This was
affecting the provider/aldente simplicity substantially for a
very specific use case.

Piping and formatting readers is very simple, and recording the output
ontop of that can be managed internally by aldente. The web ui should
strive to not impact the simplicity if possible.